Simplify Gradle cleanup
