+++ build-logic/src/main/kotlin/com/sourcegraph/buildlogic/SharedArtifacts.kt
@@ -3,7 +3,6 @@ package com.sourcegraph.buildlogic
 import org.gradle.api.Project
 import org.gradle.api.artifacts.Configuration
 import org.gradle.api.file.Directory
-import org.gradle.api.file.FileSystemLocation
 import org.gradle.api.provider.Provider
 
 /**
@@ -43,26 +42,3 @@ fun Project.publishDirectoryArtifact(
     val elements = configurations.consumable(name).get()
     artifacts.add(elements.name, directory) { builtBy(producedBy) }
 }
-
-/**
- * Builds the `kotlinc` arguments that load the scip-kotlinc compiler plugin from [pluginClasspath]
- * (the resolved shaded jar) and point it at [sourceroot]/[targetroot].
- *
- * The mapping lives here, in compiled build logic, rather than in a build script: a `.map {}`
- * lambda declared in a `.gradle.kts` file captures a hidden reference to the script object, which
- * the configuration cache cannot serialize.
- */
-fun scipKotlincPluginArgs(
-    pluginClasspath: Provider<Set<FileSystemLocation>>,
-    sourceroot: String,
-    targetroot: String,
-): Provider<List<String>> =
-    pluginClasspath.map { locations ->
-        listOf(
-            "-Xplugin=${locations.single().asFile.absolutePath}",
-            "-P",
-            "plugin:scip-kotlinc:sourceroot=$sourceroot",
-            "-P",
-            "plugin:scip-kotlinc:targetroot=$targetroot",
-        )
-    }
